Move `auth/token_exchange.go` to `http/fetch_grant_token.go`
