+++ internal/http/fetch_grant_token.go
@@ -1,4 +1,4 @@
-package auth
+package http
 
 import (
 	"bytes"
